+++ chic-admin/src/main/java/com/github/chic/admin/service/RoleService.java
@@ -6,5 +6,5 @@ import com.github.chic.entity.Role;
 import java.util.List;
 
 public interface RoleService extends IService<Role> {
-    List<Role> listByAdminId(Integer adminId);
+    List<Role> listByAdminId(Long adminId);
 }
