+++ chic-admin/src/main/java/com/github/chic/admin/util/SecurityUtils.java
@@ -19,7 +19,7 @@ public class SecurityUtils {
     /**
      * 获取当前用户ID
      */
-    public static Integer getCurrentAdminId() {
+    public static Long getCurrentAdminId() {
         return getCurrentAdmin().getId();
     }
 }
